Text to Image/requirements.txt
torch~=2.0.0
transformers~=4.28.1
diffusers~=0.15.1
Flask~=2.2.3
Pillow~=9.5.0